Refactor : Improving namespaces
Rearranging files to organize code by intended purpose, rather than by
design pattern.
