+++ Corgibytes.Freshli.Cli.Test/Functionality/GitCommitIdentifierTest.cs
@@ -1,4 +1,5 @@
 using Corgibytes.Freshli.Cli.Functionality;
+using Corgibytes.Freshli.Cli.Functionality.Git;
 using Corgibytes.Freshli.Cli.Test.Common;
 using Xunit;
 using Xunit.Abstractions;
