+++ Corgibytes.Freshli.Cli.Test/Repositories/MockCachedGitSourceRepository.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.IO;
+using Corgibytes.Freshli.Cli.DataModel;
 using Corgibytes.Freshli.Cli.Functionality;
 using Corgibytes.Freshli.Cli.Repositories;
