+++ Corgibytes.Freshli.Cli.Test/Services/MockGitArchiveProcess.cs
@@ -1,14 +1,14 @@
 using System.IO;
 using Corgibytes.Freshli.Cli.Functionality;
-using Corgibytes.Freshli.Cli.Services;
+using Corgibytes.Freshli.Cli.Functionality.Git;
 
 namespace Corgibytes.Freshli.Cli.Test.Services;
 
 public class MockGitArchiveProcess : IGitArchiveProcess
 {
-    public string Run(GitRepository gitRepository, GitCommitIdentifier gitCommitIdentifier, string gitPath, DirectoryInfo cacheDirectory)
+    public string Run(GitSource gitSource, GitCommitIdentifier gitCommitIdentifier, string gitPath, DirectoryInfo cacheDirectory)
     {
-        return "tmp/.freshli/histories/" + gitRepository.Hash + "/" + gitCommitIdentifier;
+        return "tmp/.freshli/histories/" + gitSource.Hash + "/" + gitCommitIdentifier;
     }
 }
