+++ Corgibytes.Freshli.Cli/CommandRunners/Git/GitCloneCommandRunner.cs
@@ -4,6 +4,7 @@ using System.CommandLine.IO;
 using Corgibytes.Freshli.Cli.CommandOptions;
 using Corgibytes.Freshli.Cli.Extensions;
 using Corgibytes.Freshli.Cli.Functionality;
+using Corgibytes.Freshli.Cli.Functionality.Git;
 using Corgibytes.Freshli.Lib;
 
 namespace Corgibytes.Freshli.Cli.CommandRunners.Git;
@@ -17,7 +18,7 @@ public class GitCloneCommandRunner : CommandRunner<GitCloneCommandOptions>
     public override int Run(GitCloneCommandOptions options, InvocationContext context)
     {
         // Clone or pull the given repository and branch.
-        var gitRepository = new GitRepository(options.RepoUrl, options.Branch, options.CacheDir);
+        var gitRepository = new GitSource(options.RepoUrl, options.Branch, options.CacheDir);
         try
         {
             gitRepository.CloneOrPull(options.GitPath);
