+++ Corgibytes.Freshli.Cli/Commands/MainCommand.cs
@@ -1,5 +1,6 @@
 ﻿using System.CommandLine;
 using System.IO;
+using Corgibytes.Freshli.Cli.DataModel;
 using Corgibytes.Freshli.Cli.Functionality;
 
 namespace Corgibytes.Freshli.Cli.Commands;
