+++ Corgibytes.Freshli.Cli/Functionality/Git/IGitArchiveProcess.cs
@@ -1,0 +1,8 @@
+using System.IO;
+
+namespace Corgibytes.Freshli.Cli.Functionality.Git;
+
+public interface IGitArchiveProcess
+{
+    public string Run(GitSource gitSource, GitCommitIdentifier gitCommitIdentifier, string gitPath, DirectoryInfo cacheDirectory);
+}
