--- Corgibytes.Freshli.Cli/Services/IGitArchiveProcess.cs
@@ -1,9 +1,0 @@
-using System.IO;
-using Corgibytes.Freshli.Cli.Functionality;
-
-namespace Corgibytes.Freshli.Cli.Services;
-
-public interface IGitArchiveProcess
-{
-    public string Run(GitRepository gitRepository, GitCommitIdentifier gitCommitIdentifier, string gitPath, DirectoryInfo cacheDirectory);
-}
